Refactor - separate 'single word' stuff
