+++ src/js/tiles/core/wordFreq/index.ts
@@ -102,7 +102,7 @@ export class WordFreqTile implements ITileProvider {
     }
 
     supportsQueryType(qt:QueryType, lang1:string, lang2?:string):boolean {
-        return qt === QueryType.SINGLE_QUERY;
+        return qt === QueryType.SINGLE_QUERY || qt === QueryType.CMP_QUERY;
     }
 
     disable():void {
